+++ lib/date/nowUnix.ts
@@ -4,7 +4,7 @@
  * @returns Current unix timestamp in seconds
  */
 function nowUnix ():number {
-    return Math.floor(Date.now()/1000);
+    return (Date.now()/1000) | 0;
 }
 
 export {nowUnix, nowUnix as default};