+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Improved
+- **perf**: Minor performance bump in array/sort thanks to swapping Math.floor in favor of bitwise `| 0`
+- **perf**: Minor performance bump in array/shuffle thanks to swapping Math.floor in favor of bitwise `| 0`
+- **perf**: Minor performance bump in date/nowUnix thanks to swapping Math.floor in favor of bitwise `| 0`
+- **perf**: Minor performance bump in date/nowUnixMs thanks to removing unnecessary Math.floor
+- **perf**: Minor performance bump in date/toUnix thanks to swapping Math.floor in favor of bitwise `| 0`
+- **perf**: Minor performance bump in date/format thanks to swapping Math.floor in favor of bitwise `| 0`
+- **perf**: Minor performance bump in number/randomIntBetween thanks to swapping Math.floor in favor of bitwise `| 0`
+
 ## [12.15.0] - 2024-07-16
 ### Improved
 - **dx**: array/sort - filter_fn option now makes use of a generic, ensuring that the provided filter function is correctly typed